Add links to redirect to forms
Our froms are actually hosted by google forms. We just want an easy way
to link to them. We also would prefer not to expose them to source
control. So we link to local html files and redirect for each form. Best
thing I could find that mimics envoronment variables in a ruby project.